feature: fix GPG signing of Git commits
For some clients, I don't want to sign things using my personal SSH
signing key. I want to be able to use GPG signing for $reasons.

This change ensures that I get prompted for the credentials for the key
if it's not already in gpg-agent.

See https://stackoverflow.com/questions/51504367/gpg-agent-forwarding-inappropriate-ioctl-for-device